# Improve chunked array parsing significantly