--- a/lib/parser.js
+++ b/lib/parser.js
@@ -205,18 +205,45 @@ function parseArray (parser) {
   if (length === -1) {
     return null
   }
-
   var responses = new Array(length)
-  var bufferLength = parser.buffer.length - 3
-  for (var i = 0; i < length; i++) {
+  return parseArrayElements(parser, responses, 0)
+}
+
+/**
+ * Parse chunked redis array response
+ * @param parser
+ * @returns {*}
+ */
+function parseArrayChunks (parser) {
+  return parseArrayElements(parser, parser.arrayCache, parser.arrayPos)
+}
+
+/**
+ * Parse redis array response elements
+ * @param parser
+ * @param responses
+ * @param i
+ * @returns {*}
+ */
+function parseArrayElements (parser, responses, i) {
+  var bufferLength = parser.buffer.length
+  while (i < responses.length) {
+    var offset = parser.offset
     if (parser.offset >= bufferLength) {
+      parser.arrayCache = responses
+      parser.arrayPos = i
+      parser.offset = offset
       return
     }
     var response = parseType(parser, parser.buffer[parser.offset++])
     if (response === undefined) {
+      parser.arrayCache = responses
+      parser.arrayPos = i
+      parser.offset = offset
       return
     }
     responses[i] = response
+    i++
   }
 
   return responses
@@ -297,6 +324,8 @@ function JavascriptRedisParser (options) {
   this.bigOffset = 0
   this.totalChunkSize = 0
   this.bufferCache = []
+  this.arrayCache = null
+  this.arrayPos = 0
 }
 
 /**
@@ -399,6 +428,7 @@ function concatBuffer (parser, length) {
  * @returns {undefined}
  */
 JavascriptRedisParser.prototype.execute = function execute (buffer) {
+  var arr
   if (this.buffer === null) {
     this.buffer = buffer
     this.offset = 0
@@ -410,18 +440,34 @@ JavascriptRedisParser.prototype.execute = function execute (buffer) {
     buffer.copy(newBuffer, remainingLength, 0, buffer.length)
     this.buffer = newBuffer
     this.offset = 0
+    if (this.arrayCache) {
+      arr = parseArrayChunks(this)
+      if (!arr) {
+        return
+      }
+      this.returnReply(arr)
+      this.arrayCache = null
+    }
   } else if (this.totalChunkSize + buffer.length >= this.bigStrSize) {
     this.bufferCache.push(buffer)
-    // The returned type might be Array * (42) and in that case we can't improve the parsing currently
-    if (this.optionReturnBuffers === false && this.buffer[this.offset] === 36) {
+    if (this.optionReturnBuffers === false && !this.arrayCache) {
       this.returnReply(concatBulkString(this))
       this.buffer = buffer
-    } else { // This applies for arrays too
+    } else {
       this.buffer = concatBuffer(this, this.totalChunkSize + buffer.length)
       this.offset = 0
+      if (this.arrayCache) {
+        arr = parseArrayChunks(this)
+        if (!arr) {
+          this.bigStrSize = 0
+          this.bufferCache = []
+          return
+        }
+        this.returnReply(arr)
+        this.arrayCache = null
+      }
     }
     this.bigStrSize = 0
-    this.totalChunkSize = 0
     this.bufferCache = []
   } else {
     this.bufferCache.push(buffer)
@@ -434,7 +480,9 @@ JavascriptRedisParser.prototype.execute = function execute (buffer) {
     var type = this.buffer[this.offset++]
     var response = parseType(this, type)
     if (response === undefined) {
-      this.offset = offset
+      if (!this.arrayCache) {
+        this.offset = offset
+      }
       return
     }