… benchmark/old/hiredis.js
… benchmark/old/javascript.js
… benchmark/old/parser.js