Parsing time for big arrays (e.g. 4mb+) is now linear and works well for arbitrary array sizes.
 
This case is a magnitude faster than before and almost as fast as the hiredis parser. It is the only known use case where the hiredis parser is currently till faster. In most cases the JS parser will actually be faster, since the benchmark is set to unrealistic data.
 
 OLD CODE: * bigArray x 1.09 ops/sec ±2.15% (7 runs sampled)
 OLD BUF: * bigArray x 1.23 ops/sec ±2.67% (8 runs sampled)
 
 NEW CODE: * bigArray x 273 ops/sec ±2.09% (85 runs sampled)
 NEW BUF: * bigArray x 259 ops/sec ±1.32% (85 runs sampled)

There might still be a better way to do this but at least right now I could not come up with anything better.

Fixes #23