@BridgeAR 2016-12-02
@andrasq @mscdex as you looked at the parser before I thought you might not mind to give me some input on the way big arrays behave currently.

Currently I can not find a way to optimize those further and I'm not that happy with the current implementation and thought still be some potential left. As you can see I currently cache the already read part and continue from that point on with the next chunk.